Push DEV_ tags for master container
For now, pushes in DEV will trigger the creation of a dev_ tag.
The workflow is not covering the replacement of the production tag because this should be a manual operation a few weeks/months after migration.

To be more precise, we should automate this flow after we bump buildbot version.